minor changes to models
