+++ api/utils_api.go
@@ -6,6 +6,7 @@ import (
 )
 
 func StartGame(remoteAddr string) (string, *models.Game, *models.Player) {
+	// TODO: change this to a 6-char string
 	newGameUuid := uuid.NewString()
 	newPlayerUuid := uuid.NewString()
