--- models/models_wsmsg.go
@@ -1,48 +1,0 @@
-package models
-
-const (
-	CodeStartGame = iota
-	CodeEndGame
-	CodeAttack
-	CodeReady
-)
-
-type SignalStruct struct {
-	Code int `json:"code"`
-}
-
-type StartGameResp struct {
-	GameUuid string  `json:"game_uuid"`
-	HostUuid string  `json:"host_uuid"`
-	HostGrid [][]int `json:"host_grid"`
-}
-
-// type IncomingMessage struct {
-//   ActionSignal string
-//   Payload      interface{}
-// }
-//
-// type ActionSignalStruct struct {
-// 	StartGame string
-// 	EndGame   string
-// }
-//
-// var ActionSignal = &ActionSignalStruct{
-// 	StartGame: "0",
-// 	EndGame:   "1",
-// }
-//
-// type StatusSignalStruct struct {
-// 	Success string
-// 	Failure string
-// }
-//
-// var StatusSignal = &StatusSignalStruct{
-// 	Success: "0",
-// 	Failure: "1",
-// }
-
-// type IncomingMessage struct {
-// 	ActionSignal string `json:"action_signal"`
-// 	Payload      interface{}
-// }
